Extract NSIndexSet+Bond.
